feat(app): Ensure the App works for Rhino
I'm also taking the opportunity to refactor the App structure, separating things into clear inputs, outputs and simulation.
