+++ app/requirements.txt
@@ -1,7 +1,7 @@
 streamlit_vtkjs>=0.2.0
 honeybee-vtk>=0.37.3
 pollination-streamlit>=0.2.1
-pollination-streamlit-io==0.30.7
+pollination-streamlit-io==0.31.3
 pollination-streamlit-viewer>=0.4.0
 ladybug-charts>=1.18.5
 honeybee-energy==1.91.35
